Feat[lwjgl]: add vulkan to lwjgl dlopen hook, move hook to new file
