+++ app_pojavlauncher/src/main/jni/Android.mk
@@ -46,6 +46,7 @@ LOCAL_SRC_FILES := \
     utils.c \
     stdio_is.c \
     java_exec_hooks.c \
+    lwjgl_dlopen_hook.c \
     driver_helper/nsbypass.c
 
 ifeq ($(TARGET_ARCH_ABI),arm64-v8a)
